Move classes in Substitutes.cs to the correct namespace
